Add support for edismax field aliases.
See: https://cwiki.apache.org/confluence/display/solr/The+Extended+DisMax+Query+Parser
under "Field aliasing using per-field qf overrides".

I have implemented this is in my own application and it works very nicely.

I added a test, but the tests seems to be sensitive to the order of the options. So I
modified the test harness to cast query options to sets so that they can be compared
without regard to order. Hopefullly this is okay?