Add support for facet ranges.
I'm using this in my application and it works really well.

I added a test but (like my last PR) the test harness compares sequences in
order, even when order doesn't matter. I changed the test harness to ignore
order. Please verify if this is an acceptable change.

(Alternatively, maybe it would be better to use a specific assertion:
http://stackoverflow.com/a/7473137/122763)